Split CIIP application extraction in multiple files
